Don't trust the workers.
A large and hopefully successful change in the way Ghost Town handles workers.
Instead of separate messages for "ready", "done", "cancel", and "process",
switch to one for the master and one for the workers. Workers simply notify the
master when they're done, and then the master continuously gives them work.
When pageDeath runs out, the workers die, and the master reassigns their items.
There's no longer any way for _workerQueue to get out of sync with reality.